Do not remove pid file if the process is not daemonized - fixes issue stop script will not stop the process Thanks iceman (http://sourceforge.net/tracker/?func=detail&atid=1315312&aid=3575479&group_id=312498)

+++ voipmonitor.cpp
@@ -1226,7 +1226,9 @@ int main(int argc, char *argv[]) {
 	}
 
 	free(sipportmatrix);
-	unlink(opt_pidfile);
+	if (opt_fork){
+		unlink(opt_pidfile);
+	}
 	if(opt_cachedir[0] != '\0') {
 		terminating2 = 1;
 		pthread_join(cachedir_thread, NULL);
